enable errata 430973 workaround
As mentioned at [1] u-boot unconditionally enabled a workaround
for errata 430973. Although this no longer applies for sillicon
revision r1p7 like the ccgx, it causes segfaults when enabled
in the the bootloader. Since deploying a new MLO is not that easy
turn on the workaround in linux. The code at [2] can trigger
this condition.

[1] http://lists.denx.de/pipermail/u-boot/2013-July/158377.html
[2] github.com/downloads/ssvb/ssvb.github.com/ssvb-cpuburn-a8.S