clk: imx: correct PLL2 rate calculation
As i.MX6's PLL2 also support a fractional-N synthesizer, so we need to
consider the NUM and DENOM's value to get a correct rate, as fraction
may be used in some cases.

Remove round_rate and set_rate for PLL2, as it is NOT allowed to be
changed in kernel dynamically, otherwise, PFDs and DDR may NOT work
normally, it normally should be changed in u-boot before DDR is enabled.

Signed-off-by: Anson Huang <Anson.Huang@nxp.com>