davinci_emac: fix setting the mac from DT
commit 9120bd6e9f77 ("net: davinci_emac: Get device dm816x MAC address
using the cpsw code") sets the mac address to the one stored in the chip
unconditionally, overwritten the one already set from the device tree.
This patch makes sure the mac from DT is preserved.

On a am3517 this address is incorrectly read as all zeros, making it
impossible to set a valid mac address without this patch.

Signed-off-by: Jeroen Hofstee <jhofstee@victronenergy.com>
cc: Tony Lindgren <tony@atomide.com>
cc: David S. Miller <davem@davemloft.net>
cc: Anant Gole <anantgole@ti.com>
cc: Chaithrika U S <chaithrika@ti.com>
cc: linux-omap@vger.kernel.org