ubifs: Add option to create UBI FS version 4 on empty UBI volume
Instead of creating an ubifs file system with format version 5 by
default on empty UBI volumes add a compile option to create an older ubi
with file system format version 4 instated. This allows it to be mount
as a volume on kernel versions < 4.10, which does not support format
version 5.
We saw that some people can not access their older data when they
downgraded from kernel 4.14 to kernel 4.9 to prevent this this option
would help.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>

From https://patchwork.ozlabs.org/patch/884261/