ccgx.dts: enable pullup/downs for gpio pins
Since the export of gpio pins is now done in userspace it takes a
while before the pins actual gets driven. In the meantime there are
floating, which e.g. can cause the buzzer to make noise. Prevent this
by pulling them in the right direction.

see https://github.com/victronenergy/venus/issues/270