can: ti_hecc: fix close when napi poll is active
When closing this CAN interface while napi poll is active, for example with:
`ip link set can0 down` several interfaces freeze. This seemed to be caused
by napi_disable called from ti_hecc_close expecting the scheduled probe to
either return quota or call napi_complete. Since the poll functions has a
check for netif_running it returns 0 and doesn't call napi_complete and hence
violates the napi its expectation.

So remove this check, so either napi_complete is called or quota is returned.

Signed-off-by: Jeroen Hofstee <jhofstee@victronenergy.com>

---
Also posted upstream https://patchwork.ozlabs.org/patch/895956. This is easily
reproducable on a ccgx, make sure the CAN-bus is up, flood it with CAN messages
the ccgx has to read and try to take the CAN-bus down. Most of the times the
CAN bus and ethernet userland programs are stuck in uninterruptable sleep when
doing so. Other user land processes are still scheduled.