can: ti_hecc: don't repoll but interrupt again
Besides "can: ti_hecc: fix close when napi poll is active", there is
another problem with the CAN-bus. When the CAN-bus is busy receiving
and e.g. wget downloads a big file over ethernet, both stall after a
while. While rewriting the poll routine completely and trimming it
down again to find the root cause I ended up with:

https://patchwork.ozlabs.org/patch/896387/

Although this more inline with what napi expects a driver to do, I
don't understand at the moment how it is capable to prevent taking
down the ethernet interface. It is included to check if this only
helps for my test setup or fixes things for other test setups as well.

for completeness: there is yet another problem with this driver,
receiving out of order packages when under considerable load.