ccgx: use performance as default governor
The CAN-bus can receive a burst of messages and seems to be unable to
keep up with that in ondemand mode. I haven't looked into why that is,
but it could also be that waiting for a stable clock causes the problem
instead of the clock frequency itself.

This restores the default performance governor (and hence that is not in
the commit itself, it only disables the ondemand as default)