ARM: dts: cerbo: enable usb bluetooth interface
The Bluetooth interface on the RTL8723BU was disabled since it doesn't
work.  Since then, this part has been replaced by RTL8723DU which has
working Bluetooth.  Dropping the disabling of the Bluetooth interface
allows it to be used in addition to the other on-board adapter which
has problems of its own.

On old devices, the RTL8723BU remains disabled through a USB quirks
blacklisting.