leds: lt3593: disable interrupts while pulsing gpio
If the control pin stays high or low for more than 128 us while
the pulses are being applied, the current programming starts over,
resulting in an incorrect setting.  To prevent this, interrupts
must be disabled for the duration of the loop.  The pulse width
is reduced to 250 ns, the minimum required, in order to minimise
the time with interrupts disabled.

Signed-off-by: Mans Rullgard <mans@mansr.com>