mcp251xfd: change the definition of failed
Instead of marking the driver as failed upon the first CRC communication
failure, mark it as failed when it happened for the third time in the last
minute. Set restarted if a CRC error triggers a reset of the chip.

Note: this can be squashed with: "mcp251xfd: add sysfs entries to monitor
its state", but that is already pushed.