ARM: dts: cerbo: spi2: fix assigned-clock-rates
The assigned clock rate of 600 MHz appears to be too high for the SPI
host controller in the A20. It can lead to the loss of 1..4 bytes of
the received SPI data. In some cases, a doubling of the read data
has been be observed.

The board has an mcp2518fd connected to SPI2. Its external clock is 40
MHz, which results in a maximum SPI frequency of 17 MHz. The SPI host
controller only supports even clock dividers. Using 300 MHz input
clock make no difference on the SPI clock:

input Clock	divisor	SPI Clock
---------------------------------
600 MHz		36	16.67 MHz
300 MHz		18	16.67 MHz

Dial down the clock rate to 300 MHz to fix the problem.

Signed-off-by: Marc Kleine-Budde <mkl@pengutronix.de>